+++ src/CMakeLists.txt
@@ -53,6 +53,7 @@ target_link_libraries(
     mediaelch_network
     mediaelch_renamer
     mediaelch_scrapers
+    mediaelch_scraper_scraper_apis
     mediaelch_scraper_tv_custom
     mediaelch_scraper_tv_thetvdb
     mediaelch_scraper_tv_imdb
