+++ src/scrapers/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_subdirectory(api)
 add_subdirectory(image)
 add_subdirectory(tv_show)
