+++ src/scrapers/tv_show/tmdb/CMakeLists.txt
@@ -1,7 +1,6 @@
 add_library(
   mediaelch_scraper_tv_tmdb OBJECT
   TmdbTv.cpp
-  TmdbTvApi.cpp
   TmdbTvEpisodeParser.cpp
   TmdbTvSeasonParser.cpp
   TmdbTvEpisodeScrapeJob.cpp
@@ -12,7 +11,7 @@ add_library(
 )
 
 target_link_libraries(
-  mediaelch_scraper_tv_tmdb PRIVATE Qt5::Sql Qt5::Widgets Qt5::Multimedia
-                                    Qt5::Xml
+  mediaelch_scraper_tv_tmdb PRIVATE mediaelch_scraper_scraper_apis Qt5::Sql
+                                    Qt5::Widgets Qt5::Multimedia Qt5::Xml
 )
 mediaelch_post_target_defaults(mediaelch_scraper_tv_tmdb)
